Common: split room_file.cpp to have generic read functions separate
This is mainly to allow stand-alone tools include parts of the data reading that do not bring too much other dependencies.
