+++ Common/CMakeLists.txt
@@ -77,6 +77,7 @@ target_sources(common
     game/plugininfo.h
     game/room_file.cpp
     game/room_file.h
+    game/room_file_base.cpp
     game/room_file_deprecated.cpp
     game/room_version.h
     game/roomstruct.cpp
