+++ Common/game/room_file.cpp
@@ -41,67 +41,6 @@ namespace AGS
 namespace Common
 {
 
-RoomDataSource::RoomDataSource()
-    : DataVersion(kRoomVersion_Undefined)
-{
-}
-
-String GetRoomFileErrorText(RoomFileErrorType err)
-{
-    switch (err)
-    {
-    case kRoomFileErr_NoError:
-        return "No error.";
-    case kRoomFileErr_FileOpenFailed:
-        return "Room file was not found or could not be opened.";
-    case kRoomFileErr_FormatNotSupported:
-        return "Format version not supported.";
-    case kRoomFileErr_UnexpectedEOF:
-        return "Unexpected end of file.";
-    case kRoomFileErr_UnknownBlockType:
-        return "Unknown block type.";
-    case kRoomFileErr_OldBlockNotSupported:
-        return "Block type is too old and not supported by this version of the engine.";
-    case kRoomFileErr_BlockDataOverlapping:
-        return "Block data overlapping.";
-    case kRoomFileErr_IncompatibleEngine:
-        return "This engine cannot handle requested room content.";
-    case kRoomFileErr_ScriptLoadFailed:
-        return "Script load failed.";
-    case kRoomFileErr_InconsistentData:
-        return "Inconsistent room data, or file is corrupted.";
-    case kRoomFileErr_PropertiesBlockFormat:
-        return "Unknown format of the custom properties block.";
-    case kRoomFileErr_InvalidPropertyValues:
-        return "Errors encountered when reading custom properties.";
-    case kRoomFileErr_BlockNotFound:
-        return "Required block was not found.";
-    }
-    return "Unknown error.";
-}
-
-// Read room data header and check that we support this format
-static HRoomFileError OpenRoomFileBase(Stream *in, RoomDataSource &src)
-{
-    src.DataVersion = (RoomFileVersion)in->ReadInt16();
-    if (src.DataVersion < kRoomVersion_250b || src.DataVersion > kRoomVersion_Current)
-        return new RoomFileError(kRoomFileErr_FormatNotSupported, String::FromFormat("Required format version: %d, supported %d - %d", src.DataVersion, kRoomVersion_250b, kRoomVersion_Current));
-    return HRoomFileError::None();
-}
-
-HRoomFileError OpenRoomFile(const String &filename, RoomDataSource &src)
-{
-    // Cleanup source struct
-    src = RoomDataSource();
-    // Try to open room file
-    Stream *in = File::OpenFileRead(filename);
-    if (in == nullptr)
-        return new RoomFileError(kRoomFileErr_FileOpenFailed, String::FromFormat("Filename: %s.", filename.GetCStr()));
-    src.Filename = filename;
-    src.InputStream.reset(in);
-    return OpenRoomFileBase(in, src);
-}
-
 HRoomFileError OpenRoomFileFromAsset(const String &filename, RoomDataSource &src)
 {
     // Cleanup source struct
@@ -112,49 +51,7 @@ HRoomFileError OpenRoomFileFromAsset(const String &filename, RoomDataSource &src
         return new RoomFileError(kRoomFileErr_FileOpenFailed, String::FromFormat("Filename: %s.", filename.GetCStr()));
     src.Filename = filename;
     src.InputStream.reset(in);
-    return OpenRoomFileBase(in, src);
-}
-
-
-enum RoomFileBlock
-{
-    kRoomFblk_None              = 0,
-    // Main room data
-    kRoomFblk_Main              = 1,
-    // Room script text source (was present in older room formats)
-    kRoomFblk_Script            = 2,
-    // Old versions of compiled script (no longer supported)
-    kRoomFblk_CompScript        = 3,
-    kRoomFblk_CompScript2       = 4,
-    // Names of the room objects
-    kRoomFblk_ObjectNames       = 5,
-    // Secondary room backgrounds
-    kRoomFblk_AnimBg            = 6,
-    // Contemporary compiled script
-    kRoomFblk_CompScript3       = 7,
-    // Custom properties
-    kRoomFblk_Properties        = 8,
-    // Script names of the room objects
-    kRoomFblk_ObjectScNames     = 9,
-    // End of room data tag
-    kRoomFile_EOF               = 0xFF
-};
-
-String GetRoomBlockName(RoomFileBlock id)
-{
-    switch (id)
-    {
-    case kRoomFblk_Main: return "Main";
-    case kRoomFblk_Script: return "TextScript";
-    case kRoomFblk_CompScript: return "CompScript";
-    case kRoomFblk_CompScript2: return "CompScript2";
-    case kRoomFblk_ObjectNames: return "ObjNames";
-    case kRoomFblk_AnimBg: return "AnimBg";
-    case kRoomFblk_CompScript3: return "CompScript3";
-    case kRoomFblk_Properties: return "Properties";
-    case kRoomFblk_ObjectScNames: return "ObjScNames";
-    }
-    return "unknown";
+    return ReadRoomHeader(src);
 }
 
 void ReadRoomObject(RoomObjectInfo &obj, Stream *in)
@@ -587,36 +484,6 @@ HRoomFileError ReadRoomBlock(RoomStruct *room, Stream *in, RoomFileBlock block, 
         String::FromFormat("Type: %s", ext_id.GetCStr()));
 }
 
-
-static HRoomFileError OpenNextBlock(Stream *in, RoomFileVersion data_ver, RoomFileBlock &block_id, String &ext_id, soff_t &block_len)
-{
-    // The block meta format is shared with the main game file extensions
-    //    - 1 byte - an old-style unsigned numeric ID:
-    //               where 0 would indicate following string ID,
-    //               and 0xFF indicates end of extension list.
-    //    - 16 bytes - string ID of an extension (if numeric ID is 0).
-    //    - 4 or 8 bytes - length of extension data, in bytes (size depends on format version).
-    int b = in->ReadByte();
-    if (b < 0)
-        return new RoomFileError(kRoomFileErr_UnexpectedEOF);
-
-    block_id = (RoomFileBlock)b;
-    if (block_id == kRoomFile_EOF)
-        return HRoomFileError::None(); // end of list
-
-    if (block_id > 0)
-    { // old-style block identified by a numeric id
-        ext_id = GetRoomBlockName(block_id);
-        block_len = data_ver < kRoomVersion_350 ? in->ReadInt32() : in->ReadInt64();
-    }
-    else
-    { // new style block identified by a string id
-        ext_id = String::FromStreamCount(in, 16);
-        block_len = in->ReadInt64();
-    }
-    return HRoomFileError::None();
-}
-
 HRoomFileError ReadRoomData(RoomStruct *room, Stream *in, RoomFileVersion data_ver)
 {
     room->DataVersion = data_ver;
@@ -631,7 +498,7 @@ HRoomFileError ReadRoomData(RoomStruct *room, Stream *in, RoomFileVersion data_v
         RoomFileBlock block_id;
         String ext_id;
         soff_t block_len;
-        HRoomFileError err = OpenNextBlock(in, data_ver, block_id, ext_id, block_len);
+        HRoomFileError err = OpenNextRoomBlock(in, data_ver, block_id, ext_id, block_len);
         if (!err)
             return err;
         if (ext_id.IsEmpty())
@@ -816,7 +683,7 @@ HRoomFileError ExtractScriptText(String &script, Stream *in, RoomFileVersion dat
         RoomFileBlock block_id;
         String ext_id;
         soff_t block_len;
-        HRoomFileError err = OpenNextBlock(in, data_ver, block_id, ext_id, block_len);
+        HRoomFileError err = OpenNextRoomBlock(in, data_ver, block_id, ext_id, block_len);
         if (!err)
             return err;
         if (ext_id.IsEmpty())
@@ -836,43 +703,6 @@ HRoomFileError ExtractScriptText(String &script, Stream *in, RoomFileVersion dat
         in->Seek(block_len); // skip block
     };
     return new RoomFileError(kRoomFileErr_BlockNotFound);
-}
-
-
-// Type of function that writes single room block.
-typedef void(*PfnWriteBlock)(const RoomStruct *room, Stream *out);
-// Generic function that saves a block and automatically adds its size into header
-void WriteBlock(const RoomStruct *room, RoomFileBlock block, const String &ext_id, PfnWriteBlock writer, Stream *out)
-{
-    // Write block's header
-    out->WriteByte(block);
-    if (block == kRoomFblk_None) // new-style string id
-        ext_id.WriteCount(out, 16);
-    soff_t sz_at = out->GetPosition();
-    out->WriteInt64(0); // block size placeholder
-    // Call writer to save actual block contents
-    writer(room, out);
-
-    // Now calculate the block's size...
-    soff_t end_at = out->GetPosition();
-    soff_t block_size = (end_at - sz_at) - sizeof(int64_t);
-    // ...return back and write block's size in the placeholder
-    out->Seek(sz_at, Common::kSeekBegin);
-    out->WriteInt64(block_size);
-    // ...and get back to the end of the file
-    out->Seek(0, Common::kSeekEnd);
-}
-
-// Helper for new-style blocks with string id
-void WriteBlock(const RoomStruct *room, const String &ext_id, PfnWriteBlock writer, Stream *out)
-{
-    WriteBlock(room, kRoomFblk_None, ext_id, writer, out);
-}
-
-// Helper for old-style blocks with only numeric id
-void WriteBlock(const RoomStruct *room, RoomFileBlock block, PfnWriteBlock writer, Stream *out)
-{
-    WriteBlock(room, block, String(), writer, out);
 }
 
 void WriteInteractionScripts(const InteractionScripts *interactions, Stream *out)
@@ -1025,21 +855,21 @@ HRoomFileError WriteRoomData(const RoomStruct *room, Stream *out, RoomFileVersio
     // Header
     out->WriteInt16(data_ver);
     // Main data
-    WriteBlock(room, kRoomFblk_Main, WriteMainBlock, out);
+    WriteRoomBlock(room, kRoomFblk_Main, WriteMainBlock, out);
     // Compiled script
     if (room->CompiledScript)
-        WriteBlock(room, kRoomFblk_CompScript3, WriteCompSc3Block, out);
+        WriteRoomBlock(room, kRoomFblk_CompScript3, WriteCompSc3Block, out);
     // Object names
     if (room->ObjectCount > 0)
     {
-        WriteBlock(room, kRoomFblk_ObjectNames, WriteObjNamesBlock, out);
-        WriteBlock(room, kRoomFblk_ObjectScNames, WriteObjScNamesBlock, out);
+        WriteRoomBlock(room, kRoomFblk_ObjectNames, WriteObjNamesBlock, out);
+        WriteRoomBlock(room, kRoomFblk_ObjectScNames, WriteObjScNamesBlock, out);
     }
     // Secondary background frames
     if (room->BgFrameCount > 1)
-        WriteBlock(room, kRoomFblk_AnimBg, WriteAnimBgBlock, out);
+        WriteRoomBlock(room, kRoomFblk_AnimBg, WriteAnimBgBlock, out);
     // Custom properties
-    WriteBlock(room, kRoomFblk_Properties, WritePropertiesBlock, out);
+    WriteRoomBlock(room, kRoomFblk_Properties, WritePropertiesBlock, out);
 
     // Write end of room file
     out->WriteByte(kRoomFile_EOF);
